requirements.txt

altair                    6.0.0
anywidget                 0.9.21
asttokens                 3.0.1
attrs                     25.4.0
blinker                   1.9.0
cachetools                6.2.6
certifi                   2026.1.4
charset-normalizer        3.4.4
click                     8.3.1
colorama                  0.4.6
comm                      0.2.3
debugpy                   1.8.20
decorator                 5.2.1
executing                 2.2.1
gitdb                     4.0.12
GitPython                 3.1.46
greenlet                  3.3.1
idna                      3.11
ipykernel                 7.2.0
ipython                   9.10.0
ipython_pygments_lexers   1.1.1
ipywidgets                8.1.8
jedi                      0.19.2
Jinja2                    3.1.6
joblib                    1.5.3
jsonschema                4.26.0
jsonschema-specifications 2025.9.1
jupyter_client            8.8.0
jupyter_core              5.9.1
jupyterlab_widgets        3.0.16
MarkupSafe                3.0.3
matplotlib-inline         0.2.1
narwhals                  2.16.0
nest-asyncio              1.6.0
numpy                     2.4.2
packaging                 26.0
pandas                    2.3.3
parso                     0.8.5
pillow                    12.1.0
pip                       26.0.1
platformdirs              4.5.1
plotly                    6.5.2
prompt_toolkit            3.0.52
protobuf                  6.33.5
psutil                    7.2.2
psycopg2-binary           2.9.11
psygnal                   0.15.1
pure_eval                 0.2.3
pyarrow                   23.0.0
pydeck                    0.9.1
Pygments                  2.19.2
python-dateutil           2.9.0.post0
python-dotenv             1.2.1
pytz                      2025.2
pyzmq                     27.1.0
referencing               0.37.0
requests                  2.32.5
rpds-py                   0.30.0
scikit-learn              1.8.0
scipy                     1.17.0
six                       1.17.0
smmap                     5.0.2
SQLAlchemy                2.0.46
stack-data                0.6.3
streamlit                 1.54.0
streamlit-autorefresh     1.0.1
tenacity                  9.1.3
threadpoolctl             3.6.0
toml                      0.10.2
tornado                   6.5.4
traitlets                 5.14.3
typing_extensions         4.15.0
tzdata                    2025.3
urllib3                   2.6.3
watchdog                  6.0.0
wcwidth                   0.6.0
widgetsnbextension        4.0.15
